kindex: DevC++ (OpenGL WIN), Visual Studio (OpenGL WIN, SDL) WIN: keyboard, mouse
git-svn-id: svn://localhost/steel.new/trunk@68 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/code/physic/physic_engine.cpp
@@ -12,7 +12,7 @@ bool PhysicEngine::init(std::string _conf)
 
 	conf = (Config*)res->get(Res::config, _conf);
 
-	g.z = -conf->getf("g", 0.0);
+	g.z = -(float)conf->getf("g", 0.0);
 	g.x = 0.0f;
 	g.y = 0.0f;
